sliding window(development ongoing)
