+++ src/read_pcap_dpkt.py
@@ -8,358 +8,20 @@ import csv
 import sys
 import collections
 import threading
-from threading import Thread
+from threading import Thread, local
 import time
 import os
+from IpStat import IpStat
+import pdb
 
 window_size = 10
+window_interval = 1
+
+
+
 filename='./data/21Feb_pcap.pcap'
 #top_num = 3
 
-
-class IpStat:
-    def __init__(self, target_IPs) -> None:
-        self.local_stat = {}
-        self.local_stat["total incoming traffic by packet"] = 0
-        self.local_stat["total incoming traffic by byte"] = 0
-        self.local_stat["total outgoing traffic by packet"] = 0
-        self.local_stat["total outgoing traffic by byte"] = 0
-        self.target_IPs = target_IPs
-
-    def update_stat(self, size, eth_src, eth_dst, ip_src, ip_dst, proto, port_src, port_dst):
-        
-        for ip in self.target_IPs:
-
-            if (ip_src == ip or ip_dst == ip) and self.local_stat.get(ip) == None:
-                self.local_stat[ip] = {}
-                
-                self.local_stat[ip]["incoming traffic"] = {}
-                self.local_stat[ip]["incoming traffic"]["#packet"] = 0
-                self.local_stat[ip]["incoming traffic"]["traffic in bytes"] = 0
-                self.local_stat[ip]["outgoing traffic"] = {}
-                self.local_stat[ip]["outgoing traffic"]["#packet"] = 0
-                self.local_stat[ip]["outgoing traffic"]["traffic in bytes"] = 0
-                self.local_stat[ip]["external IPs"] = {}
-                self.local_stat[ip]["internal ports"] = {}
-                self.local_stat[ip]["external ports"] = {}
-                self.local_stat[ip]["protocols"] = {}
-
-
-            if ip_src == ip:
-                # outgoing packet
-                self.local_stat["total outgoing traffic by packet"] += 1
-                self.local_stat["total outgoing traffic by byte"] += size
-                
-                self.local_stat[ip]["outgoing traffic"]["#packet"] += 1
-                self.local_stat[ip]["outgoing traffic"]["traffic in bytes"] += size
-                
-
-
-                if self.local_stat[ip]["external IPs"].get(ip_dst) == None:
-                    self.local_stat[ip]["external IPs"][ip_dst] = {}
-                    self.local_stat[ip]["external IPs"][ip_dst]["#incoming packet"] = 0
-                    self.local_stat[ip]["external IPs"][ip_dst]["incoming traffic in bytes"] = 0
-                    self.local_stat[ip]["external IPs"][ip_dst]["#outgoing packet"] = 0
-                    self.local_stat[ip]["external IPs"][ip_dst]["outgoing traffic in bytes"] = 0
-                
-                if self.local_stat[ip]["internal ports"].get(port_src) == None:
-                    self.local_stat[ip]["internal ports"][port_src] = {}
-                    self.local_stat[ip]["internal ports"][port_src]["#packet"] = 0
-                    self.local_stat[ip]["internal ports"][port_src]["traffic in bytes"] = 0
-
-                if self.local_stat[ip]["external ports"].get(port_dst) == None:
-                    self.local_stat[ip]["external ports"][port_dst] = {}
-                    self.local_stat[ip]["external ports"][port_dst]["#packet"] = 0
-                    self.local_stat[ip]["external ports"][port_dst]["traffic in bytes"] = 0
-
-                if self.local_stat[ip]["protocols"].get(proto) == None:
-                    self.local_stat[ip]["protocols"][proto] = {}
-                    self.local_stat[ip]["protocols"][proto]["#packet"] = 0
-                    self.local_stat[ip]["protocols"][proto]["traffic in bytes"] = 0
-
-                # update info
-                self.local_stat[ip]["external IPs"][ip_dst]["#outgoing packet"] += 1
-                self.local_stat[ip]["external IPs"][ip_dst]["outgoing traffic in bytes"] += size
-
-                self.local_stat[ip]["internal ports"][port_src]["#packet"] += 1
-                self.local_stat[ip]["internal ports"][port_src]["traffic in bytes"] += size
-                self.local_stat[ip]["external ports"][port_dst]["#packet"] += 1
-                self.local_stat[ip]["external ports"][port_dst]["traffic in bytes"] += size
-
-                self.local_stat[ip]["protocols"][proto]["#packet"] += 1
-                self.local_stat[ip]["protocols"][proto]["traffic in bytes"] += size
-
-
-            elif ip_dst == ip:
-                # incoming packet
-
-                self.local_stat["total incoming traffic by packet"] += 1
-                self.local_stat["total incoming traffic by byte"] += size
-
-                self.local_stat[ip]["incoming traffic"]["#packet"] += 1
-                self.local_stat[ip]["incoming traffic"]["traffic in bytes"] += size
-                
-                if self.local_stat[ip]["external IPs"].get(ip_src) == None:
-                    self.local_stat[ip]["external IPs"][ip_src] = {}
-                    self.local_stat[ip]["external IPs"][ip_src]["#incoming packet"] = 0
-                    self.local_stat[ip]["external IPs"][ip_src]["incoming traffic in bytes"] = 0
-                    self.local_stat[ip]["external IPs"][ip_src]["#outgoing packet"] = 0
-                    self.local_stat[ip]["external IPs"][ip_src]["outgoing traffic in bytes"] = 0
-
-                if self.local_stat[ip]["internal ports"].get(port_dst) == None:
-                    self.local_stat[ip]["internal ports"][port_dst] = {}
-                    self.local_stat[ip]["internal ports"][port_dst]["#packet"] = 0
-                    self.local_stat[ip]["internal ports"][port_dst]["traffic in bytes"] = 0
-
-                if self.local_stat[ip]["external ports"].get(port_src) == None:
-                    self.local_stat[ip]["external ports"][port_src] = {}
-                    self.local_stat[ip]["external ports"][port_src]["#packet"] = 0
-                    self.local_stat[ip]["external ports"][port_src]["traffic in bytes"] = 0
-                
-                if self.local_stat[ip]["protocols"].get(proto) == None:
-                    self.local_stat[ip]["protocols"][proto] = {}
-                    self.local_stat[ip]["protocols"][proto]["#packet"] = 0
-                    self.local_stat[ip]["protocols"][proto]["traffic in bytes"] = 0
-
-                # update info
-                self.local_stat[ip]["external IPs"][ip_src]["#incoming packet"] += 1
-                self.local_stat[ip]["external IPs"][ip_src]["incoming traffic in bytes"] += size
-
-                self.local_stat[ip]["internal ports"][port_dst]["#packet"] += 1
-                self.local_stat[ip]["internal ports"][port_dst]["traffic in bytes"] += size
-                self.local_stat[ip]["external ports"][port_src]["#packet"] += 1
-                self.local_stat[ip]["external ports"][port_src]["traffic in bytes"] += size
-
-                self.local_stat[ip]["protocols"][proto]["#packet"] += 1
-                self.local_stat[ip]["protocols"][proto]["traffic in bytes"] += size    
-
-    '''
-        Analyze local database and generate features
-        return a 1 dimension dictionary
-    '''
-    # [ip, start_time, end_time, in rate, out rate, in bite, out bite, avg in size, avg out size, 
-    # \ top ext ip by pkt, top ext ip by size, total ex IP, top in port(pkt/%/byte/%), top out port(pkt/%/byte/%), top proto(pkt/byte)]
-    def analyze_features(self):
-        out_list = []
-        i = 0
-        for ip in target_IPs:
-            # create empty row
-            out_list.append([])
-
-            # start to generate IP specific start
-            out_list[i].append(ip)
-            out_list[i].append(target_IP_types[ip])
-            out_list[i].append(currTime)
-            out_list[i].append(currTime + window_size)
-            if self.local_stat.get(ip) == None:
-                out_list[i].extend([0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,0,0,0,0,0 ])
-                i += 1 
-                continue
-
-        # packet rate(incoming/outgoing/bidirection)
-
-            if self.local_stat["total incoming traffic by packet"] != 0:
-                out_list[i].append(self.local_stat[ip]["incoming traffic"]["#packet"]/\
-                            self.local_stat["total incoming traffic by packet"])
-            else :
-                out_list[i].append(0)
-
-            if self.local_stat["total outgoing traffic by packet"] != 0:
-                out_list[i].append(self.local_stat[ip]["outgoing traffic"]["#packet"]/\
-                                self.local_stat["total outgoing traffic by packet"])
-            else:
-                out_list[i].append(0)
-        
-        # byte rate(incoming/outgoing/bidirection)
-            if self.local_stat["total incoming traffic by byte"] != 0:
-                out_list[i].append(self.local_stat[ip]["incoming traffic"]["traffic in bytes"]/\
-                                self.local_stat["total incoming traffic by byte"])
-            else:
-                out_list[i].append(0)
-            
-            if self.local_stat["total outgoing traffic by byte"] != 0:
-                out_list[i].append(self.local_stat[ip]["outgoing traffic"]["traffic in bytes"]/\
-                            self.local_stat["total outgoing traffic by byte"])
-            else:
-                out_list[i].append(0)
-        
-        # average packet size(incoming/outgoing)
-            if out_list[i][-2] > 0:
-                avg_in_size = self.local_stat[ip]["incoming traffic"]["traffic in bytes"]\
-                            /self.local_stat[ip]["incoming traffic"]["#packet"]
-            else:
-                avg_in_size = 0
-
-            if out_list[i][-1] > 0:
-                avg_out_size = self.local_stat[ip]["outgoing traffic"]["traffic in bytes"]\
-                            /self.local_stat[ip]["outgoing traffic"]["#packet"]
-            else:
-                avg_out_size = 0
-
-            out_list[i].append(avg_in_size)
-            out_list[i].append(avg_out_size)
-
-        # external IPs
-            ext_ip_list = list(self.local_stat[ip]["external IPs"].keys())
-            max_pkt = 0
-            max_size = 0
-            total_pkt = 0
-            total_size = 0
-            ind = 0
-            ex_ip_pkt = 'None'
-            ex_ip_size = 'None'
-            if len(ext_ip_list) > 0:
-                for ext_ip in ext_ip_list:
-                    pkt_n = self.local_stat[ip]["external IPs"][ext_ip]["#incoming packet"] + \
-                            self.local_stat[ip]["external IPs"][ext_ip]["#outgoing packet"]
-                    size_pkt = self.local_stat[ip]["external IPs"][ext_ip]["incoming traffic in bytes"] + \
-                                self.local_stat[ip]["external IPs"][ext_ip]["outgoing traffic in bytes"]
-                    total_pkt += pkt_n
-                    total_size += size_pkt
-                    if pkt_n > max_pkt:
-                        max_pkt = pkt_n
-
-                    if size_pkt > max_size:
-                        max_size = size_pkt
-                    ind += 1
-                if total_pkt > 0:
-                    ex_ip_pkt = max_pkt/total_pkt
-                else :
-                    ex_ip_pkt = 0
-                
-                if total_size > 0:
-                    ex_ip_size = max_size/total_size
-                else:
-                    ex_ip_size = 0
-            
-            out_list[i].append(ex_ip_pkt)
-            out_list[i].append(ex_ip_size)
-            ext_ip_list = list(self.local_stat[ip]["external IPs"].keys())
-            out_list[i].append(len(ext_ip_list))
-        # ranked ports(internal/external)
-            internal_port_list = list(self.local_stat[ip]["internal ports"].keys())
-            max_pkt = 0
-            max_size = 0
-            total_pkt = 0
-            total_size = 0
-            max_ind_pkt = 0
-            max_ind_size = 0
-            ind = 0
-            out_pkt = 'None'
-            out_size = 'None'
-            if len(internal_port_list) > 0:
-                for port_n in internal_port_list:
-                    pkt_n = self.local_stat[ip]["internal ports"][port_n]["#packet"]
-                    size_pkt = self.local_stat[ip]["internal ports"][port_n]["traffic in bytes"]
-                    total_pkt += pkt_n
-                    total_size += size_pkt
-                    if pkt_n > max_pkt:
-                        max_pkt = pkt_n
-                        max_ind_pkt = ind
-
-                    if size_pkt > max_size:
-                        max_size = size_pkt
-                        max_ind_size = ind
-                    ind += 1
-                out_pkt = internal_port_list[max_ind_pkt]
-                out_size = internal_port_list[max_ind_size]
-            
-            if total_pkt > 0:
-                max_pkt = max_pkt/total_pkt
-            
-            if total_size > 0:
-                max_size = max_size/total_size
-
-
-            out_list[i].append(out_pkt)
-            out_list[i].append(max_pkt)
-            out_list[i].append(out_size)
-            out_list[i].append(max_size)
-
-            external_port_list = list(self.local_stat[ip]["external ports"].keys())
-            max_pkt = 0
-            max_size = 0
-            total_pkt = 0
-            total_size = 0
-            max_ind_pkt = 0
-            max_ind_size = 0
-            ind = 0
-            out_pkt = 'None'
-            out_size = 'None'
-            if len(external_port_list) > 0:
-                for port_n in external_port_list:
-                    pkt_n = self.local_stat[ip]["external ports"][port_n]["#packet"]
-                    size_pkt = self.local_stat[ip]["external ports"][port_n]["traffic in bytes"]
-                    total_pkt += pkt_n
-                    total_size += size_pkt
-                    if pkt_n > max_pkt:
-                        max_pkt = pkt_n
-                        max_ind_pkt = ind
-
-                    if size_pkt > max_size:
-                        max_size = size_pkt
-                        max_ind_size = ind
-                    ind += 1
-                out_pkt = external_port_list[max_ind_pkt]
-                out_size = external_port_list[max_ind_size]
-            if total_pkt > 0:
-                max_pkt = max_pkt/total_pkt
-            
-            if total_size > 0:
-                max_size = max_size/total_size
-            
-            out_list[i].append(out_pkt)
-            out_list[i].append(max_pkt)
-            out_list[i].append(out_size)
-            out_list[i].append(max_size)
-        
-        # protocol(bidirection)
-            proto_list = list(self.local_stat[ip]["protocols"].keys())
-            max_pkt = 0
-            max_size = 0
-            total_pkt = 0
-            total_size = 0
-            max_ind_pkt = 0
-            max_ind_size = 0
-            ind = 0
-            out_pkt = 'None'
-            out_size = 'None'
-            if len(proto_list) > 0:
-                for pro in proto_list:
-                    pkt_n = self.local_stat[ip]["protocols"][pro]["#packet"]
-                    size_pkt = self.local_stat[ip]["protocols"][pro]["traffic in bytes"]
-                    total_pkt += pkt_n
-                    total_size += size_pkt
-                    if pkt_n > max_pkt:
-                        max_pkt = pkt_n
-                        max_ind_pkt = ind
-
-                    if size_pkt > max_size:
-                        max_size = size_pkt
-                        max_ind_size = ind
-                    ind += 1
-                out_pkt = proto_list[max_ind_pkt]
-                out_pkt_per = max_pkt/total_pkt
-                out_size = proto_list[max_ind_size]
-                out_size_per = max_size/total_size
-            
-            #out_list[i].append(out_pkt)
-            out_list[i].append(out_pkt_per)
-            #out_list[i].append(out_size)
-            out_list[i].append(out_size_per)
-
-            i += 1
-
-        return out_list
-
-    def clear(self):
-        """Create an empty data structure for holding IP stats"""
-        self.local_stat = {}
-        self.local_stat["total incoming traffic by packet"] = 0
-        self.local_stat["total incoming traffic by byte"] = 0
-        self.local_stat["total outgoing traffic by packet"] = 0
-        self.local_stat["total outgoing traffic by byte"] = 0
-        
 
 
 def mac_addr(address):
@@ -414,9 +76,9 @@ def plot_piechart(dic, chartName, threshold):
 def write_csv(out_list):
     for output in out_list:
         writer.writerow(output)
-        print("hellos")
 
-
+def aggregate_window(sub_windows_list):
+    None
 
 # def get_attr_for_ip(ip_addr):
 #     """Get a pre-defined set of attributes for the given IP address
@@ -583,6 +245,7 @@ pcap = dpkt.pcap.Reader(f)
 
 monitor_window = 60   # size of the time window in seconds, according to the pcap timestamp
 
+# read target IPs and their types
 target_IP_types = {}
 ip_read = open('./data/Host-ShortList.csv', 'r', encoding='UTF8', newline='')
 ip_reader = csv.reader(ip_read)
@@ -594,48 +257,6 @@ for row in ip_reader:
     target_IP_types[row[0]] = row[2]
 ip_read.close()
 target_IPs = list(target_IP_types.keys())
-# ips under monitoring
-'''target_ips = ["129.94.0.197",\
-    "129.94.0.196",\
-    "129.94.0.193",\
-    "129.94.0.192",\
-    "149.171.124.33",\
-    "149.171.144.58",\
-    "149.171.197.177",\
-    "129.94.52.190",\
-    "149.171.144.50",\
-    "149.171.194.50",\
-    "149.171.184.237",\
-    "149.171.155.194",\
-    "149.171.105.163",\
-    "129.94.164.21",\
-    "149.171.207.19",\
-    "149.171.99.112",\
-    "149.171.99.18",\
-    "149.171.194.17",\
-    "149.171.193.32",\
-    "129.94.8.226",\
-    "129.94.8.238",\
-    "129.94.8.163",\
-    "129.94.8.160",\
-    "129.94.8.77",\
-    "129.94.8.81",\
-    "129.94.8.173",\
-    "129.94.8.237",\
-    "149.171.12.3",\
-    "129.94.254.213",\
-    "149.171.66.213",\
-    "149.171.0.34",\
-    "149.171.133.31",\
-    "149.171.148.27",\
-    "129.94.183.252",\
-    "149.171.168.194",\
-    "149.171.29.35",\
-    "149.171.99.123",\
-    "149.171.99.99",\
-    "149.171.99.94",\
-    "149.171.99.185"\
-    ]'''
 
 p_id = None
 p_time = None
@@ -648,21 +269,35 @@ p_proto = None
 p_port_src = None
 p_port_dst = None
 
-# create local da'tabase
-local_database = IpStat(target_IPs)
+# the local database is essentially is list of maximum length window_size/window_interval
+# e.g. a window size of 60s and window interval of 10s give maximum length of 6
+local_database = []
+
+window_IpStats = None
+
+# create the first entry of the local database(from time 0 - window_size)
+
+#local_database.append(window1_stat)
 
 #header = ['Packet ID', 'TIME', 'Size', 'eth.src', 'eth.dst', 'IP.src', 'IP.dst', 'IP.proto', 'port.src', 'port.dst']
 
-currTime = None
+curr_time = None
+accumulated_window = 0
+curr_stat = None
 
-#count = 0'
+
 for timestamp, buf in pcap:
-    if currTime == None:
-        currTime = int(timestamp)
+    if curr_time == None:
+        curr_time = int(timestamp)
 
-    if (int(timestamp) < (currTime + window_size)):
-    
+    if not curr_stat:
+        curr_stat = IpStat(target_IPs, curr_time, curr_time + window_interval)
+        #local_database.append(window1_stat)
 
+
+    if (int(timestamp) < (curr_time + window_interval)):
+        
+ 
         eth = dpkt.ethernet.Ethernet(buf)
 
         if not isinstance(eth.data, dpkt.ip.IP):
@@ -673,6 +308,8 @@ for timestamp, buf in pcap:
         do_not_fragment = bool(ip.off & dpkt.ip.IP_DF)
         more_fragments = bool(ip.off & dpkt.ip.IP_MF)
         fragment_offset = ip.off & dpkt.ip.IP_OFFMASK
+
+        link_layer_protocol = eth.get_type(eth.type).__name__
 
         srcport = 0
         dstport = 0
@@ -713,14 +350,40 @@ for timestamp, buf in pcap:
         p_port_src = srcport
         p_port_dst = dstport
 
-        # update local database according to this packet
-        local_database.update_stat(p_size, p_eth_src, p_eth_dst, p_ip_src, p_ip_dst, p_proto, p_port_src, p_port_dst)
+
+        # update the current entry of the local database with this packet's stat
+        curr_stat.update_stat(p_size, p_eth_src, p_eth_dst, p_ip_src, p_ip_dst, p_proto, p_port_src, p_port_dst)
     
     else:
+        # append currrent stat to local database
+        local_database.append(curr_stat)
+        # window is full
+        if len(local_database) == 6:
+            # first time get full
+            window_IpStats = local_database[0]
+            for stat in local_database[1:]:
+                window_IpStats += stat
+            print("finish first round")
+            write_csv(window_IpStats.analyze_features(target_IPs, target_IP_types))
+            pdb.set_trace()
+                    
+        elif len(local_database) == 7:
+            if not window_IpStats:
+                raise Exception("something went wrong")
+            # remove the oldest entry, add the latest
+            popped = local_database.pop(0)
+            window_IpStats -= popped
+            #print("popped {} to {}".format(popped.start_time, popped.end_time))
+            window_IpStats += curr_stat
+            #print("added {} to {}".format(curr_stat.start_time, curr_stat.end_time))
+            write_csv(window_IpStats.analyze_features(target_IPs, target_IP_types))
+            pdb.set_trace()
+        curr_time += window_interval
+
+        # create a new entry
+        curr_stat = IpStat(target_IPs, curr_time, curr_time + window_interval)
         
-        write_csv(local_database.analyze_features())
-        local_database.clear()
-        currTime += window_size
+
         
 f.close()
 csvf.close()
